feat: рефакторинг провайдеров и стора
