+++ src/entities/offers/model/selectors.ts
@@ -1,4 +1,4 @@
-import {State} from '../../../shared/api';
+import {State} from '../../../providers/StoreProvider/store';
 
 export const getOffers = (state: State) => state.offersList;
 export const getOffersCount = (state: State): number => state.offersCount.countOffers;
