+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { App } from './app';
-import {store} from './providers/store-provider.tsx';
-import {fetchOffersListAction} from './entities/offers/model/action.ts';
+import {store} from './providers/StoreProvider/store';
+import {fetchOffersListAction} from './entities/offers/model';
 
 store.dispatch(fetchOffersListAction());
