+++ src/providers/RouterProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './router-provider.tsx';
