+++ src/providers/StoreProvider/constants/constants.ts
@@ -1,0 +1,12 @@
+export enum APIRoute {
+  Offers = '/offers',
+  Favorites = '/favorite',
+  Comments = '/comments',
+  Auth = '/login'
+}
+
+// export enum NameSpace {
+//   Offers = 'OFFERS',
+//   Comment = 'COMMENT',
+//   User = 'USER',
+// }
